requirements.txt
fastapi
uvicorn
prisma
pandas
pydantic
coverage
python-dotenv
unittest-xml-reporting